remove old `operations` module, rename `operations2` -> `operations`
