+++ src/configureStore.ts
@@ -7,8 +7,7 @@ import promiseMiddleware from "redux-promise-middleware";
 import thunkMiddleware from "redux-thunk";
 
 import reducers from "./reducers";
-import { actionCanceler } from "reducers/operations";
-import { transactionsTracker, successDismisser } from "reducers/operations2";
+import { transactionsTracker, successDismisser } from "reducers/operations";
 import { persistStore, createTransform } from 'redux-persist';
 
 export const history = createHistory();
@@ -17,7 +16,6 @@ const store = createStore(
   reducers,
   composeWithDevTools(   // makes the store available to the Chrome redux dev tools
     applyMiddleware(
-      actionCanceler(),
       transactionsTracker,
       successDismisser(),
       thunkMiddleware,
